ya se arreglo alv, BD

+++ senseor/database/migrations/2020_03_07_063127_create_casas_table.php
@@ -16,7 +16,7 @@ class CreateCasasTable extends Migration
         Schema::create('casas', function (Blueprint $table) {
             $table->bigIncrements('clave_casa')->unsigned();
             $table->char('Nombre_casas',100);
-            $table->bigIncrements('privadas')->unsigned();
+            $table->BigInteger('privadas')->unsigned();
             $table->integer('Numero de casa');
             $table->foreign('privadas')->references('clave_privada')->on('privadas');
             $table->timestamps();
